refactor: remove estimate fee and transaction type
