+++ src/account/interface.ts
@@ -30,22 +30,6 @@ export abstract class AccountInterface extends ProviderInterface {
   public abstract signer: SignerInterface;
 
   /**
-   * @deprecated Use estimateInvokeFee or estimateDeclareFee instead
-   * Estimate Fee for executing an INVOKE transaction on starknet
-   *
-   * @param calls the invocation object containing:
-   * - contractAddress - the address of the contract
-   * - entrypoint - the entrypoint of the contract
-   * - calldata - (defaults to []) the calldata
-   *
-   * @returns response from estimate_fee
-   */
-  public abstract estimateFee(
-    calls: AllowArray<Call>,
-    estimateFeeDetails?: EstimateFeeDetails
-  ): Promise<EstimateFeeResponse>;
-
-  /**
    * Estimate Fee for executing an INVOKE transaction on starknet
    *
    * @param calls the invocation object containing:
